requirements.txt
discord.py>=2.3,<3
fastapi>=0.110,<1
uvicorn>=0.29,<1
python-dotenv>=1.0,<2
